requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
numpy>=1.24.0
openai>=1.0.0
tiktoken>=0.5.0
pinecone-client>=3.0.0
google-generativeai>=0.5.0
anthropic>=0.25.0
cohere>=5.0.0
voyageai>=0.2.0
